+++ src/pages/Search/components/TitleText/TitleText.module.scss
