+++ src/pages/Search/components/TitleText/TitleText.tsx
